♻️ refactor(reflect): move plugins/reflect/ to internal/reflect/
Reflect is a core background service, not a swappable plugin. Move to
internal/ while keeping plugin host registration intact for lifecycle
management, scheduler integration, and admin API.
