+++ cmd/stella/plugins_imports.go
@@ -13,7 +13,7 @@ import (
 	_ "github.com/CherryHQ/stella/plugins/providers/openai-response"
 
 	// Plugin runtimes.
-	_ "github.com/CherryHQ/stella/plugins/reflect"
+	_ "github.com/CherryHQ/stella/internal/reflect"
 
 	// Plugin tools.
 	_ "github.com/CherryHQ/stella/plugins/tools/webfetch"
